requirements.txt
gspread
pandas
google.oauth2
google.auth.transport.requests